+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added ability to ask for how many plays someone else has on a server.
+
 ## [v6.0.8](https://github.com/lexicalunit/spellbot/releases/tag/v6.0.8) - 2021-05-20
 
 ### Changed